requirements.txt
alembic==1.8.0
attrs==21.4.0
click==8.1.3
coverage==6.4.1
cycler==0.11.0
Flask==2.1.2
Flask-Migrate==3.1.0
Flask-Moment==1.0.2
Flask-SQLAlchemy==2.5.1
fonttools==4.33.3
greenlet==1.1.2
importlib-metadata==4.11.4
importlib-resources==5.7.1
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.3
Mako==1.2.0
MarkupSafe==2.1.1
matplotlib==3.5.2
numpy==1.22.4
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.1.1
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
pytest-cov==3.0.0
python-dateutil==2.8.2
python-dotenv==0.20.0
six==1.16.0
SQLAlchemy==1.4.37
tomli==2.0.1
Werkzeug==2.0.1
yapf==0.32.0
zipp==3.8.0
